requirements/base.txt
# Core
django==5.2.2
djangorestframework==3.16.1
django-environ==0.12.0

# Database
psycopg2-binary==2.9.11

# API & Utils
django-filter==25.2
django-cors-headers==4.9.0
inflection==0.5.1
pyyaml==6.0.3
requests==2.31.0

# Auth
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0

# Tasks & Cache
celery==5.6.0
redis==7.1.0